+++ desktop/applis/apps/core/games/belote/src/main/java/cards/belote/HandBelote.java
@@ -146,7 +146,7 @@ public final class HandBelote implements Iterable<CardBelote> {
     @return la carte aleatoire choisie*/
     CardBelote tirerUneCarteAleatoire(AbstractGenerator _gene) {
         //      return jouer((int)Math.floor(total()*MonteCarlo.randomDouble()));
-        return jouer((int) MonteCarloUtil.randomLong(total(),_gene));
+        return jouer(MonteCarloUtil.randomLong(total(),_gene));
         //0<=total()*Math.random()<total()
         //Donc 0<=Math.floor(total()*Math.random())<Math.floor(total())=total()
     }
