+++ desktop/applis/apps/core/games/president/src/main/java/cards/president/HandPresident.java
@@ -195,7 +195,7 @@ public final class HandPresident implements Iterable<CardPresident> {
     @return la carte aleatoire choisie*/
     CardPresident tirerUneCarteAleatoire(AbstractGenerator _gene) {
 //        return jouer((int)Math.floor(total()*MonteCarlo.randomDouble()));
-        return jouer((int) MonteCarloUtil.randomLong(total(),_gene));
+        return jouer(MonteCarloUtil.randomLong(total(),_gene));
         //0<=total()*Math.random()<total()
         //Donc 0<=Math.floor(total()*Math.random())<Math.floor(total())=total()
     }
