+++ desktop/applis/apps/core/games/solitaire/src/main/java/cards/solitaire/HandSolitaire.java
@@ -118,7 +118,7 @@ public final class HandSolitaire implements Iterable<CardSolitaire> {
      @return la carte aleatoire choisie*/
     CardSolitaire tirerUneCarteAleatoire(AbstractGenerator _gene) {
 //        return jouer((int)Math.floor(total()*MonteCarlo.randomDouble()));
-        return jouer((int) MonteCarloUtil.randomLong(total(),_gene));
+        return jouer(MonteCarloUtil.randomLong(total(),_gene));
         //0<=total()*Math.random()<total()
         //Donc 0<=Math.floor(total()*Math.random())<Math.floor(total())=total()
     }
