+++ desktop/applis/apps/gui/gui_games/cardgamesgui/src/main/java/cards/gui/dialogs/EditorBelote.java
@@ -225,7 +225,7 @@ public final class EditorBelote extends DialogBelote implements SetterSelectedCa
         int donneur_ = editorCards.getListe().getSelectedIndex();
         if (donneur_ == nombreDeJoueurs_) {
 //          donneur_=(byte)Math.floor(nombreDeJoueurs_*MonteCarlo.randomDouble());
-            donneur_=(int)MonteCarloUtil.randomLong(nombreDeJoueurs_,getFrames().getGenerator());
+            donneur_= MonteCarloUtil.randomLong(nombreDeJoueurs_,getFrames().getGenerator());
         }
         DealBelote donne_=new DealBelote(mains_,donneur_);
         partie = new GameBelote(GameType.EDIT,donne_,getReglesBelote());
