+++ desktop/applis/apps/gui/gui_games/cardgamesgui/src/main/java/cards/gui/dialogs/EditorPresident.java
@@ -225,7 +225,7 @@ public final class EditorPresident extends DialogPresident implements SetterSele
         int donneur_ = editorCards.getListe().getSelectedIndex();
         if (donneur_ == nombreDeJoueurs_) {
 //            donneur_=(byte)Math.floor(nombreDeJoueurs_*MonteCarlo.randomDouble());
-            donneur_=(int)MonteCarloUtil.randomLong(nombreDeJoueurs_,getFrames().getGenerator());
+            donneur_= MonteCarloUtil.randomLong(nombreDeJoueurs_,getFrames().getGenerator());
         }
         DealPresident donne_=new DealPresident(mains_,donneur_);
         partie = new GamePresident(GameType.EDIT,donne_,getReglesPresident(), new Ints());
